requirements.txt
asgiref==3.6.0
Django==4.2
django-ckeditor==6.5.1
django-cors-headers==3.14.0
django-jazzmin==2.6.0
django-js-asset==2.0.0
django-rest-framework==0.1.0
django-rest-passwordreset==1.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary==2.9.6
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.3
tzdata==2023.3
